requirements.txt
XlsxWriter==3.0.1
xlrd==2.0.1
xgboost==1.5.2
tqdm==4.64.0
statsmodels==0.12.2
sklearn-pandas==2.2.0
shap==0.34.0
scipy==1.5.1
scikit-learn==0.24.1
scikit-learn-extra==0.2.0
pytz==2022.1
pyod==0.8.7
pandas==1.1.4
openpyxl==2.5.7
numpy==1.20.0
joblib==1.2.0
ipython==8.12.2
ipykernel==6.23.1
imbalanced-learn==0.7.0
icalendar==4.1.0
holidays==0.19
hdbscan==0.8.31
dill==0.3.3
Cython==0.29.33
cloudpickle==2.0.0
streamlit
mpld3